requirements.txt
# Telegram Bot
python-telegram-bot==22.5

# FastAPI Backend
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.17.2

# S3 Yandex Object Storage
boto3==1.42.9
PyJWT==2.10.1

# Video Processing
yt-dlp==2025.12.8
ffmpeg-python==0.2.0

# Async Tasks
celery[redis]==5.6.0
redis==7.1.0
flower==2.0.1

# Yandex Cloud SDK
yandexcloud==0.371.0

# Image Processing
Pillow==12.0.0

# Document Processing
python-docx==1.1.2
pdf2image==1.17.0
pypdf==5.1.0

# Data Validation
pydantic==2.12.5
pydantic-settings==2.12.0

# HTTP
httpx==0.28.1
requests==2.32.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Other
python-dotenv==1.2.1